Allow prefix to be a whitespace/comma separated list of prefixes
This project enables pushing the same image to multiple tags, and with
this feature it can enable pushing the same image to multiple
destinations with multiple tags.

A use case for this is if we want to push to both Docker Hub and Quay.io
for example, then prefix could be `"user/repo,quay.io/user/repo"`.